refactor: remove automation client from `api/clients`
